Refactor recipients component
