+++ src/modules/capitalFlow/dialogs/setupStrategyDialog/setupStrategyDialogDistributionRecipients/index.ts
@@ -1,0 +1,2 @@
+export { SetupStrategyDialogDistributionRecipients } from './setupStrategyDialogDistributionRecipients';
+export type { ISetupStrategyDialogDistributionRecipientsProps } from './setupStrategyDialogDistributionRecipients';
